// src/api/fees/api.ts
export interface CreateFeeRequest {
    /**
     * 
     * @type {string}
     * @memberof CreateFeeRequest
     */
    'name': string;
    /**
     * 
     * @type {number}
     * @memberof CreateFeeRequest
     */
    'rate': number;
    /**
     * 
     * @type {string}
     * @memberof CreateFeeRequest
     */
    'type': CreateFeeRequestTypeEnum;
}

export const CreateFeeRequestTypeEnum = {
    Discount: 'DISCOUNT',
    Tax: 'TAX'
} as const;

export type CreateFeeRequestTypeEnum = typeof CreateFeeRequestTypeEnum[keyof typeof CreateFeeRequestTypeEnum];

/**
 * 
 * @export
 * @interface ErrorResponse
 */
export interface ErrorResponse {
    /**
     * 
     * @type {string}
     * @memberof ErrorResponse
     */
    'error'?: string;
    /**
     * 
     * @type {number}
     * @memberof ErrorResponse
     */
    'code'?: number;
}
/**
 * 
 * @export
 * @interface FeeResponse
 */
export interface FeeResponse {
    /**
     * 
     * @type {string}
     * @memberof FeeResponse
     */
    'id'?: string;
    /**
     * 
     * @type {string}
     * @memberof FeeResponse
     */
    'name'?: string;
    /**
     * 
     * @type {number}
     * @memberof FeeResponse
     */
    'rate'?: number;
    /**
     * 
     * @type {string}
     * @memberof FeeResponse
     */
    'type'?: FeeResponseTypeEnum;
    /**
     * 
     * @type {string}
     * @memberof FeeResponse
     */
    'createdAt'?: string;
}

export const FeeResponseTypeEnum = {
    Discount: 'DISCOUNT',
    Tax: 'TAX'
} as const;

export type FeeResponseTypeEnum = typeof FeeResponseTypeEnum[keyof typeof FeeResponseTypeEnum];

/**
 * 
 * @export
 * @interface GetAllFeesPaginatedResponse
 */
export interface GetAllFeesPaginatedResponse {
    /**
     * 
     * @type {Pagination}
     * @memberof GetAllFeesPaginatedResponse
     */
    'pagination'?: Pagination;
    /**
     * 
     * @type {Array<FeeResponse>}
     * @memberof GetAllFeesPaginatedResponse
     */
    'data'?: Array<FeeResponse>;
}
/**
 * 
 * @export
 * @interface Pagination
 */
export interface Pagination {
    /**
     * 
     * @type {number}
     * @memberof Pagination
     */
    'totalItems'?: number;
    /**
     * 
     * @type {number}
     * @memberof Pagination
     */
    'pageSize'?: number;
    /**
     * 
     * @type {number}
     * @memberof Pagination
     */
    'currentPage'?: number;
    /**
     * 
     * @type {number}
     * @memberof Pagination
     */
    'totalPages'?: number;
}
// ...
export interface FeesApiCreateFeeRequest {
    /**
     * 
     * @type {CreateFeeRequest}
     * @memberof FeesApiCreateFee
     */
    readonly createFeeRequest: CreateFeeRequest
}

/**
 * Request parameters for getAllFees operation in FeesApi.
 * @export
 * @interface FeesApiGetAllFeesRequest
 */
export interface FeesApiGetAllFeesRequest {
    /**
     * The starting index of the fees.
     * @type {number}
     * @memberof FeesApiGetAllFees
     */
    readonly from?: number

    /**
     * The number of fees to return.
     * @type {number}
     * @memberof FeesApiGetAllFees
     */
    readonly size?: number

    /**
     * Filter criteria for fees.
     * @type {string}
     * @memberof FeesApiGetAllFees
     */
    readonly filter?: string

    /**
     * Field to order fees by.
     * @type {'createdAt'}
     * @memberof FeesApiGetAllFees
     */
    readonly orderBy?: GetAllFeesOrderByEnum

    /**
     * Sorting direction.
     * @type {'ASC' | 'DESC'}
     * @memberof FeesApiGetAllFees
     */
    readonly direction?: GetAllFeesDirectionEnum

    /**
     * Type of fee to filter by.
     * @type {'DISCOUNT' | 'TAX'}
     * @memberof FeesApiGetAllFees
     */
    readonly type?: GetAllFeesTypeEnum
}

/**
 * FeesApi - object-oriented interface
 * @export
 * @class FeesApi
 * @extends {BaseAPI}
 */
export class FeesApi extends BaseAPI {
    /**
     * 
     * @summary Create a new fee
     * @param {FeesApiCreateFeeRequest} requestParameters Request parameters.
     * @param {*} [options] Override http request option.
     * @throws {RequiredError}
     * @memberof FeesApi
     */
    public createFee(requestParameters: FeesApiCreateFeeRequest, options?: RawAxiosRequestConfig) {
        return FeesApiFp(this.configuration).createFee(requestParameters.createFeeRequest, options).then((request) => request(this.axios, this.basePath));
    }

    /**
     * 
     * @summary Retrieve a list of fees
     * @param {FeesApiGetAllFeesRequest} requestParameters Request parameters.
     * @param {*} [options] Override http request option.
     * @throws {RequiredError}
     * @memberof FeesApi
     */
    public getAllFees(requestParameters: FeesApiGetAllFeesRequest = {}, options?: RawAxiosRequestConfig) {
        return FeesApiFp(this.configuration).getAllFees(requestParameters.from, requestParameters.size, requestParameters.filter, requestParameters.orderBy, requestParameters.direction, requestParameters.type, options).then((request) => request(this.axios, this.basePath));
    }
}

/**
 * @export
 */
export const GetAllFeesOrderByEnum = {
    CreatedAt: 'createdAt'
} as const;
export type GetAllFeesOrderByEnum = typeof GetAllFeesOrderByEnum[keyof typeof GetAllFeesOrderByEnum];
/**
 * @export
 */
export const GetAllFeesDirectionEnum = {
    Asc: 'ASC',
    Desc: 'DESC'
} as const;
export type GetAllFeesDirectionEnum = typeof GetAllFeesDirectionEnum[keyof typeof GetAllFeesDirectionEnum];
/**
 * @export
 */
export const GetAllFeesTypeEnum = {
    Discount: 'DISCOUNT',
    Tax: 'TAX'
} as const;
export type GetAllFeesTypeEnum = typeof GetAllFeesTypeEnum[keyof typeof GetAllFeesTypeEnum];